ResultsView --> ResultsData; ResultsContainer --> Results; create resultsCardRenderers config; move types around.
