+++ src/shared/components/results/Results.tsx
@@ -1,20 +1,15 @@
 import { FC } from 'react';
 
-import ResultsView from './ResultsView';
+import ResultsData from './ResultsData';
 import ResultsFacets from './ResultsFacets';
 
 import SideBarLayout from '../layouts/SideBarLayout';
 
 import './styles/results-table.scss';
 
-export enum ViewMode {
-  TABLE,
-  CARD,
-}
-
 const Results: FC = () => (
   <SideBarLayout sidebar={<ResultsFacets />}>
-    <ResultsView />
+    <ResultsData />
   </SideBarLayout>
 );
