+++ src/shared/hooks/useNSQuery.ts
@@ -9,7 +9,7 @@ import { getAPIQueryUrl } from '../config/apiUrls';
 import fieldsForUniProtKBCards from '../../uniprotkb/config/UniProtKBCardConfiguration';
 import { Column, nsToDefaultColumns } from '../config/columns';
 
-import { ViewMode } from '../components/results/ResultsContainer';
+import { ViewMode } from '../components/results/ResultsData';
 import { Namespace } from '../types/namespaces';
 
 const useNSQuery = ({
